+++ enterprise/install-scripts/generic/openshift.sh
@@ -513,7 +513,7 @@
 #CONF_ROUTING_PLUGIN_USER=routinginfo
 #CONF_ROUTING_PLUGIN_PASS=...
 #   Default: <randomized>
-#   Default with CONF_NO_SCRAMBLE: routinginfopassword
+#   Default with CONF_NO_SCRAMBLE: routinginfopasswd
 
 # Settings for configuring Kerberos as user authentication method
 #
@@ -965,12 +965,12 @@ assign_pass()
 {
  # If the ENV variable is set, use it
   if [ -n "${!3}" ]; then
-    eval $1=\${!3}
+    printf -v "$1" '%s' "${!3}"
   elif is_true "$no_scramble" ; then
-    eval $1=\$2
+    printf -v "$1" '%s' "$2"
   else
     randomized=$(openssl rand -base64 20)
-    eval $1=\${randomized//[![:alnum:]]}
+    printf -v "$1" '%s' "${randomized//[![:alnum:]]}"
   fi
   passwords[$1]="${!1}"
 }
@@ -980,7 +980,7 @@ display_passwords()
   set +x
   for k in "${!passwords[@]}"; do
     out_string=""
-    matching_var=""
+    matchingvar=""
     for postfix in password password1 pass; do
       vprefix=${k%${postfix}}
       [ "$vprefix" != "$k" ] && break
@@ -1106,10 +1106,10 @@ configure_ose_yum_repos()
   for repo in infra node jbosseap_cartridge client_tools; do
     if [ "$ose_repo_base" != "" ]; then
       layout=puddle; [ -n "$CONF_CDN_LAYOUT" ] && layout=cdn
-      eval "need_${repo}_repo && def_ose_yum_repo $ose_repo_base $layout $repo"
+      "need_${repo}_repo" && def_ose_yum_repo "$ose_repo_base" "$layout" "$repo"
     fi
     if [ "$ose_extra_repo_base" != "" ]; then
-      eval "need_${repo}_repo && def_ose_yum_repo $ose_extra_repo_base extra $repo"
+      "need_${repo}_repo" && def_ose_yum_repo "$ose_extra_repo_base" 'extra' "$repo"
     fi
   done
 }
@@ -1212,7 +1212,7 @@ configure_cart_repos()
 
   local repo
   for repo in "${!url[@]}"; do
-    eval "need_${repo}_repo" || continue
+    "need_${repo}_repo" || continue
     cat <<YUM > "/etc/yum.repos.d/${repo}.repo"
 [${repo}]
 name=${repo}
@@ -1248,7 +1248,7 @@ configure_extra_repos()
 
   local repo
   for repo in "${!priority[@]}"; do
-    local url=$(eval echo '${'$repo'}')
+    local url="${!repo}"
     if [ "${url}x" != "x" ]; then
       cat <<YUM >> "${extra_repo_file}"
 [${repo}]
@@ -1294,13 +1294,13 @@ configure_rhn_channels()
 {
   if [ "x$CONF_RHN_REG_ACTKEY" != x ]; then
     echo "OpenShift: Register to RHN Classic using an activation key"
-    eval "rhnreg_ks --force --activationkey=${CONF_RHN_REG_ACTKEY} --profilename='$profile_name' ${CONF_RHN_REG_OPTS}" || abort_install
+    rhnreg_ks --force "--activationkey=${CONF_RHN_REG_ACTKEY}" "--profilename=$profile_name" ${CONF_RHN_REG_OPTS} || abort_install
   elif [[ "${CONF_RHN_USER}" && "${CONF_RHN_PASS}" ]]
   then
     echo "OpenShift: Register to RHN Classic with username and password"
     set +x # don't log password
-    echo "rhnreg_ks --force --profilename='$profile_name' --username '${CONF_RHN_USER}' ${CONF_RHN_REG_OPTS}"
-    eval "rhnreg_ks --force --profilename='$profile_name' --username '${CONF_RHN_USER}' --password '${CONF_RHN_PASS}' ${CONF_RHN_REG_OPTS}" || abort_install
+    echo "rhnreg_ks --force \"--profilename=$profile_name\" --username \"${CONF_RHN_USER}\" ${CONF_RHN_REG_OPTS}"
+    rhnreg_ks --force "--profilename=$profile_name" --username "${CONF_RHN_USER}" --password "${CONF_RHN_PASS}" ${CONF_RHN_REG_OPTS} || abort_install
     set -x
   else
     echo "OpenShift: No credentials given for RHN Classic; assuming already configured"
@@ -1335,8 +1335,8 @@ configure_rhsm_channels()
   then
     set +x # don't log password
     echo "OpenShift: Register with RHSM"
-    echo "subscription-manager register --force --username='$CONF_RHN_USER' --name '$profile_name' ${CONF_RHN_REG_OPTS}"
-    eval "subscription-manager register --force --username='$CONF_RHN_USER' --password='$CONF_RHN_PASS' --name '$profile_name' ${CONF_RHN_REG_OPTS}" || abort_install
+    echo "subscription-manager register --force \"--username=$CONF_RHN_USER\" --name \"$profile_name\" ${CONF_RHN_REG_OPTS}"
+    subscription-manager register --force "--username=$CONF_RHN_USER" "--password=$CONF_RHN_PASS" --name "$profile_name" ${CONF_RHN_REG_OPTS} || abort_install
     set -x
   else
     echo "OpenShift: No credentials given for RHSM; assuming already configured"
@@ -3217,11 +3217,7 @@ echo_installation_intentions()
 {
   echo "The following components should be installed:"
   for component in $components
-  do
-    if eval $component
-    then
-      printf '\t%s.\n' $component
-    fi
+  do "$component" && printf '\t%s.\n' "$component"
   done
 
   echo "Configuring with broker with hostname ${broker_hostname}."
@@ -3257,7 +3253,7 @@ parse_args()
       (*=*) val="${word#*\=}" ;;
       (*) val=true ;;
     esac
-    eval "CONF_${key^^}"'="$val"'
+    printf -v "CONF_${key^^}" '%s' "$val"
   done
 }
 
@@ -3410,7 +3406,7 @@ declare -A valid_settings=( [CONF_ABORT_ON_UNRECOGNIZED_SETTINGS]= [CONF_ACTIONS
   installing_something=0
   for component in $components
   do
-    if eval $component
+    if "$component"
     then
       installing_something=1
       break
@@ -3679,7 +3675,7 @@ declare -A valid_settings=( [CONF_ABORT_ON_UNRECOGNIZED_SETTINGS]= [CONF_ACTIONS
 
   # auth info for the topic from the sample routing SPI plugin
   routing_plugin_user="${CONF_ROUTING_PLUGIN_USER:-routinginfo}"
-  assign_pass routing_plugin_pass "routinginfopassword" CONF_ROUTING_PLUGIN_PASS
+  assign_pass routing_plugin_pass "routinginfopasswd" CONF_ROUTING_PLUGIN_PASS
 
   # cartridge dependency metapackages
   metapkgs="${CONF_METAPKGS:-recommended}"
@@ -3855,8 +3851,7 @@ configure_firewall()
 {
   # Disable lokkit.
   conf='/etc/sysconfig/system-config-firewall'
-  [[ "$(< "$conf")" != --disabled ]] &&
-    mv "$conf" "${conf}.bak"
+  [[ -e "$conf" && "$(< "$conf")" != --disabled ]] && mv "$conf" "${conf}.bak"
   echo '--disabled' > "$conf"
 
   # Configure iptables.
@@ -3941,7 +3936,6 @@ configure_openshift()
   node && update_openshift_facts_on_node
 
   node && broker && fix_broker_routing
-  node && install_rsync_pub_key
 
   configure_firewall_add_rules
   node && configure_gear_isolation_firewall
@@ -4065,6 +4059,8 @@ post_deploy()
 
     configure_districts
   fi
+
+  node && install_rsync_pub_key
 
   echo "OpenShift: Completed post deployment steps."
 }
